Silence some C4267 warnings "possible loss of data".

+++ src/SubPic/SubPicAllocatorPresenterImpl.cpp
@@ -1,6 +1,6 @@
 /*
  * (C) 2003-2006 Gabest
- * (C) 2006-2013 see Authors.txt
+ * (C) 2006-2014 see Authors.txt
  *
  * This file is part of MPC-HC.
  *
@@ -317,7 +317,7 @@ STDMETHODIMP CSubPicAllocatorPresenterImpl::GetString(LPCSTR field, LPWSTR* valu
     }
 
     if (!ret.IsEmpty()) {
-        size_t len = ret.GetLength();
+        int len = ret.GetLength();
         size_t sz = (len + 1) * sizeof(WCHAR);
         LPWSTR buf = (LPWSTR)LocalAlloc(LPTR, sz);
